refactor: open edit dialog 분리
