+++ src/widgets/comment/edit-dialog/index.tsx
@@ -1,14 +1,32 @@
+import { useMutation } from "@tanstack/react-query"
+
 import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Textarea } from "@/shared/ui"
 import { useDialogStore } from "@/features/dialog/model/store"
 import { useSelectedComment } from "@/entity/comment/model/selectedComment"
-import { useEditComment } from "./model/useEditComment"
+import { useEditComment } from "@/features/comment/edit/model/useEditComment"
+import { editComment } from "@/entity/comment/api/editComment"
 
 export const EDIT_COMMENT_DIALOG = "EDIT_COMMENT_DIALOG"
 
 export const EditCommentDialog = () => {
-  const { isDialogOpen, setDialogOpen } = useDialogStore()
-  const { selectedComment } = useSelectedComment()
-  const { body, setBody, mutation } = useEditComment()
+  const { isDialogOpen, setDialogOpen, closeDialog } = useDialogStore()
+  const { selectedComment, setSelectedComment } = useSelectedComment()
+  const { body, setBody } = useEditComment()
+
+  const mutation = useMutation({
+    mutationFn: (commentId: number) => {
+      return editComment(commentId, body)
+    },
+    onSuccess: (data) => {
+      setSelectedComment(null)
+      closeDialog(EDIT_COMMENT_DIALOG)
+      // 댓글 목록 수정
+      console.log(data)
+    },
+    onError: (error) => {
+      console.error("댓글 수정 오류:", error)
+    },
+  })
 
   if (!selectedComment) return null
